add function classes, question, question bank, exam

+++ src/main/java/com/example/quizeonline5/sevice/impl/ClassServiceImpl.java
@@ -1,14 +1,18 @@
 package com.example.quizeonline5.sevice.impl;
 
 
+import com.example.quizeonline5.dto.ApiResponse;
 import com.example.quizeonline5.dto.ClassDto;
+import com.example.quizeonline5.dto.StudentDto;
 import com.example.quizeonline5.entity.Classes;
 import com.example.quizeonline5.entity.Subject;
 import com.example.quizeonline5.entity.User;
+import com.example.quizeonline5.exception.ResourceNotFoundException;
 import com.example.quizeonline5.repository.ClassRepository;
 import com.example.quizeonline5.repository.SubjectRepository;
 import com.example.quizeonline5.repository.UserRepository;
 import com.example.quizeonline5.sevice.ClassService;
+import com.example.quizeonline5.utils.AppConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -28,7 +32,7 @@ public class ClassServiceImpl implements ClassService {
     private SubjectRepository subjectRepository;
 
     @Override
-    public Long createClass(ClassDto classDto) {
+    public ApiResponse createClass(ClassDto classDto) {
 
         if (classDto.getTeacherId() == null) {
             throw new IllegalArgumentException("Teacher ID must not be null");
@@ -36,11 +40,10 @@ public class ClassServiceImpl implements ClassService {
         if (classDto.getSubjectId() == null) {
             throw new IllegalArgumentException("Subject ID must not be null");
         }
-
-        User teacher = userRepository.findByUserId(classDto.getTeacherId())
+        User teacher = userRepository.findById(classDto.getTeacherId())
                 .orElseThrow(() -> new IllegalArgumentException("Teacher not found"));
 
-        Subject subject = subjectRepository.findById(classDto.getSubjectId())
+        Subject subject = subjectRepository.findById(1L)
                 .orElseThrow(() -> new IllegalArgumentException("Subject not found"));
 
         Classes classEntity = new Classes();
@@ -48,16 +51,27 @@ public class ClassServiceImpl implements ClassService {
         classEntity.setTeacher(teacher);
         classEntity.setSubject(subject);
         classEntity.setCreatedAt(LocalDateTime.now());
-        return classRepository.save(classEntity).getClassId();
+        classEntity.setUpdatedAt(LocalDateTime.now());
+        Long id =  classRepository.save(classEntity).getClassId();
+        if (id != null) {
+            return new ApiResponse(true, "Class created successfully with ID: " + id);
+        } else {
+            return new ApiResponse(false, "Class not created");
+        }
     }
 
     @Override
-    public void deleteClass(Long classId) {
+    public ApiResponse deleteClass(Long classId) {
+        Classes classes = classRepository.findById(classId).orElseThrow(() -> new ResourceNotFoundException(AppConstants.CLASSES, AppConstants.ID, classId));
+        if (classes == null) {
+            return new ApiResponse(Boolean.FALSE, "Class not found");
+        }
         classRepository.deleteById(classId);
+        return new ApiResponse(Boolean.TRUE, "Class deleted successfully");
     }
 
     @Override
-    public void updateClass(Long classId, ClassDto classDto) {
+    public ApiResponse updateClass(Long classId, ClassDto classDto) {
         Classes classEntity = classRepository.findById(classId)
                 .orElseThrow(() -> new IllegalArgumentException("Class not found"));
 
@@ -68,16 +82,17 @@ public class ClassServiceImpl implements ClassService {
         classEntity.setTeacher(teacher);
         classEntity.setUpdatedAt(java.time.LocalDateTime.now());
         classRepository.save(classEntity);
+        return new ApiResponse(Boolean.TRUE, "Update class success");
     }
 
     @Override
     public List<Map<String, Object>> getAllClasses() {
         return classRepository.findAll().stream().map(classEntity -> {
             Map<String, Object> map = new HashMap<>();
-            map.put("classId", classEntity.getClassId());
-            map.put("className", classEntity.getClassName());
-            map.put("subjectName", classEntity.getSubject().getSubjectName());
-            map.put("teacherName", classEntity.getTeacher().getFullName());
+            map.put("class_id", classEntity.getClassId());
+            map.put("class_name", classEntity.getClassName());
+            map.put("subject_name", classEntity.getSubject().getSubjectName());
+            map.put("teacher_name", classEntity.getTeacher().getFullName());
             return map;
         }).collect(Collectors.toList());
     }
@@ -89,9 +104,9 @@ public class ClassServiceImpl implements ClassService {
 
         return classEntity.getStudents().stream().map(classStudent -> {
             Map<String, Object> map = new HashMap<>();
-            map.put("studentId", classStudent.getStudent().getUserId());
-            map.put("fullName", classStudent.getStudent().getFullName());
-            map.put("email", classStudent.getStudent().getEmail());
+            map.put("student_id", classStudent.getUserId());
+            map.put("full_name", classStudent.getFullName());
+            map.put("email", classStudent.getEmail());
             return map;
         }).collect(Collectors.toList());
     }
@@ -103,16 +118,37 @@ public class ClassServiceImpl implements ClassService {
 
         User teacher = classEntity.getTeacher();
         return Map.of(
-                "teacherId", teacher.getUserId(),
-                "fullName", teacher.getFullName(),
+                "teacher_id", teacher.getUserId(),
+                "full_name", teacher.getFullName(),
                 "email", teacher.getEmail(),
-                "subjectName", classEntity.getSubject().getSubjectName()
+                "subject_name", classEntity.getSubject().getSubjectName()
         );
     }
 
     @Override
     public List<Map<String, Object>> getClassesByTeacher(Long teacherId) {
         return List.of();
+    }
+
+    @Override
+    public Classes getClassById(Long classId) {
+        return  classRepository.findById(classId)
+                .orElseThrow(() -> new IllegalArgumentException("Class not found"));
+
+    }
+
+    @Override
+    public ApiResponse addStudentToClass(Long classId, StudentDto studentDto) {
+        Classes classEntity = classRepository.findById(classId)
+                .orElseThrow(() -> new ResourceNotFoundException(AppConstants.CLASSES, AppConstants.ID, classId));
+
+        User student = userRepository.findById(studentDto.getStudentId())
+                .orElseThrow(() -> new ResourceNotFoundException(AppConstants.USER, AppConstants.ID, studentDto.getStudentId()));
+
+        classEntity.getStudents().add(student);
+        classRepository.save(classEntity);
+
+        return new ApiResponse(true, "Student added to class successfully");
     }
 
 //    @Override
